requirement.txt
Django
Pillow
djangorestframework
django-filter
---------------
django-cors-headers
djangorestframework-simplejwt
django-bootstrap-modal-forms
django-widget-tweaks
